Added isUpToDate
I added a method to my persistenceFile class to check it's date modified
against the file on the disk. I kinda wonder if we need a file class
like this or should just use the built in file class and just create
utility class that has functions like this one.