Search methods now return a Set
The search methods now return a set rather than try to add the files to
the UI directly. Seems like a much better idea design-wise and gets the
data a step closer to where it needs to be. It's at least in the same
class now. I updated the methods and the search button. Still haven't
tried to display it yet.