Filled in the constructor for the persistenceFile class
Trying to move toward being to store data so made a class we can store
the data we need about files in. I also made the fields private so we
can use getters and setter methods but I haven't added those yet. And
I'll change the other classes too to use them. I may use only getters
and require the use of the constructor with arguments.